Match source and test files
