+++ Sources/Player/Time.swift
@@ -7,17 +7,6 @@
 import AVFoundation
 
 enum Time {
-    static func timeRange(for item: AVPlayerItem?) -> CMTimeRange {
-        guard let item else {
-            return .invalid
-        }
-        guard let firstRange = item.seekableTimeRanges.first?.timeRangeValue,
-              let lastRange = item.seekableTimeRanges.last?.timeRangeValue else {
-            return !item.loadedTimeRanges.isEmpty ? .zero : .invalid
-        }
-        return CMTimeRangeFromTimeToTime(start: firstRange.start, end: lastRange.end)
-    }
-
     static func progress(for time: CMTime, in range: CMTimeRange) -> Float {
         guard range.isValid && !range.isEmpty else { return 0 }
         let elapsedTime = CMTimeGetSeconds(CMTimeSubtract(time, range.start))
