+++ Tests/PlayerTests/PeriodicTimePublisherTests.swift
@@ -1,0 +1,62 @@
+//
+//  Copyright (c) SRG SSR. All rights reserved.
+//
+//  License information is available from the LICENSE file.
+//
+
+@testable import Player
+
+import AVFoundation
+import Combine
+import Circumspect
+import XCTest
+
+final class PeriodicTimePublisherTests: XCTestCase {
+    func testPeriodicTimeDuringPlayback() throws {
+        let item = AVPlayerItem(url: TestStreams.onDemandUrl)
+        let player = AVPlayer(playerItem: item)
+        player.play()
+        try expectPublished(
+            values: [
+                CMTimeMake(value: 0, timescale: 2),
+                CMTimeMake(value: 1, timescale: 2),
+                CMTimeMake(value: 2, timescale: 2),
+                CMTimeMake(value: 3, timescale: 2),
+                CMTimeMake(value: 4, timescale: 2),
+                CMTimeMake(value: 5, timescale: 2)
+            ],
+            from: Publishers.PeriodicTimePublisher(for: player, interval: CMTimeMake(value: 1, timescale: 2)),
+            to: beClose(within: 0.5)
+        )
+    }
+
+    func testPeriodicTimeDuringSeek() throws {
+        let item = AVPlayerItem(url: TestStreams.onDemandUrl)
+        let player = AVPlayer(playerItem: item)
+
+        // TODO: Wait for the item to be ready
+
+        Task {
+            await player.seek(
+                to: CMTime(value: 3, timescale: 2),
+                toleranceBefore: .zero,
+                toleranceAfter: .zero
+            )
+        }
+        try expectPublished(
+            values: [
+                CMTimeMake(value: 3, timescale: 2),
+                CMTimeMake(value: 4, timescale: 2),
+                CMTimeMake(value: 5, timescale: 2),
+                CMTimeMake(value: 6, timescale: 2)
+            ],
+            from: Publishers.PeriodicTimePublisher(for: player, interval: CMTimeMake(value: 1, timescale: 2)),
+            to: beClose(within: 0.5)
+        )
+    }
+
+    // TODO:
+    //  - Test without playing (no events; requires a way to check that values are never emitted)
+    //  - Test with pause
+    //  - etc.
+}
