--- Tests/PlayerTests/TestNotifications.swift
@@ -1,11 +1,0 @@
-//
-//  Copyright (c) SRG SSR. All rights reserved.
-//
-//  License information is available from the LICENSE file.
-//
-
-import Foundation
-
-extension Notification.Name {
-    static let testNotification = Notification.Name("TestNotification")
-}
