--- Tests/PlayerTests/TestTypes.swift
@@ -1,20 +1,0 @@
-//
-//  Copyright (c) SRG SSR. All rights reserved.
-//
-//  License information is available from the LICENSE file.
-//
-
-import Foundation
-
-// swiftlint:disable file_types_order
-
-final class TestNSObject: NSObject {
-}
-
-final class TestObject {
-}
-
-struct TestStruct {
-}
-
-// swiftlint:enable file_types_order
